Brought up to date with JUCE 6 API changes and C++17 compiler restrictions
Just a few minor changes to get this code to compile and run with JUCE 6 and C++17. The main thing is that we have to cast the std::atomic<float>* parameters to (float*) in a few cases, but after careful examination of the C++ standard spec (the source specs; not on-line discussions at StackOverflow and the like), it became clear that this is a safe and correct change. I ran an intensive debugging session to verify the results anyway. Note that the Maximilian directory structure is not as expected by this project, but I'm not yet sure if I'll bother making a fork of that repository, unless I see that I also had to fix compiler errors in that library.